+++ packages/necroparser/test/fixtures/jsdoc.ts
@@ -1,0 +1,10 @@
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+interface Interface {
+	/**
+	 * Property
+	 *
+	 * @default 'foo'
+	 * @readonly
+	 */
+	readonly property: string;
+}