+++ packages/necroparser/test/fixtures/type-kind.ts
@@ -1,0 +1,9 @@
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+interface Interface {
+	primitive: string;
+	array: string[];
+	object: {
+		[key: string]: unknown;
+	};
+	union: 'foo' | 'bar';
+}