+++ packages/necroparser/unbuild.config.ts
@@ -1,0 +1,11 @@
+import { defineBuildConfig } from 'unbuild';
+
+export default defineBuildConfig({
+	entries: ['src/index'],
+	outDir: 'dist',
+	clean: true,
+	declaration: true,
+	rollup: {
+		emitCJS: true
+	}
+});