+++ packages/skeleton-react/src/lib/components/ProgressRing/types.ts
@@ -3,6 +3,8 @@ import * as progress from '@zag-js/progress';
 export interface ProgressRingProps extends React.PropsWithChildren, Omit<progress.Context, 'id'> {
 	/** Set the text for the scalable label */
 	label?: string;
+	/** When enabled, show a text label with the percentage amount */
+	showLabel?: boolean;
 	/** Set the stroke size (ex: 15px) */
 	strokeWidth?: string;
 	/** Defines the shape of the meter */