+++ packages/skeleton-svelte/src/lib/components/Accordion/Accordion.svelte
@@ -62,5 +62,5 @@
 <!-- @component Divide content into collapsible sections. -->
 
 <div class="{base} {padding} {spaceY} {rounded} {width} {classes}" {...api.getRootProps()} data-testid="accordion">
-	{@render children()}
+	{@render children?.()}
 </div>