+++ packages/skeleton-svelte/src/lib/components/Modal/Modal.svelte
@@ -7,6 +7,9 @@
 
 	let {
 		open = $bindable(false),
+		// Base
+		base = '',
+		classes = '',
 		// Trigger
 		triggerBase = '',
 		triggerBackground = '',
@@ -21,6 +24,7 @@
 		positionerJustify = 'justify-center',
 		positionerAlign = 'items-center',
 		positionerPadding = 'p-4',
+		positionerZIndex = '',
 		positionerClasses = '',
 		// Content
 		contentBase = '',
@@ -42,14 +46,15 @@
 	const [snapshot, send] = useMachine(
 		dialog.machine({
 			id: useId(),
-			open,
-			onOpenChange(details) {
-				open = details.open;
-			}
+			open
 		}),
 		{
 			context: {
 				...zagProps,
+				onOpenChange(details) {
+					zagProps.onOpenChange?.(details);
+					open = details.open;
+				},
 				get open() {
 					return $state.snapshot(open);
 				}
@@ -59,11 +64,13 @@
 	const api = $derived(dialog.connect(snapshot, send, normalizeProps));
 </script>
 
-<span>
+<span class="{base} {classes}" data-testid="modal">
 	<!-- Trigger -->
-	<button {...api.getTriggerProps()} class="{triggerBase} {triggerBackground} {triggerClasses}">
-		{@render trigger?.()}
-	</button>
+	{#if trigger}
+		<button {...api.getTriggerProps()} class="{triggerBase} {triggerBackground} {triggerClasses}">
+			{@render trigger()}
+		</button>
+	{/if}
 	{#if api.open}
 		<!-- Backdrop -->
 		<div
@@ -77,7 +84,7 @@
 		<div
 			use:portal
 			{...api.getPositionerProps()}
-			class="{positionerBase} {positionerDisplay} {positionerJustify} {positionerAlign} {positionerPadding} {positionerClasses}"
+			class="{positionerBase} {positionerDisplay} {positionerJustify} {positionerAlign} {positionerPadding} {positionerZIndex} {positionerClasses}"
 			in:fly={transitionsPositionerIn}
 			out:fly={transitionsPositionerOut}
 		>