+++ packages/skeleton-svelte/src/lib/components/ProgressRing/ProgressRing.svelte
@@ -6,6 +6,7 @@
 
 	let {
 		label,
+		showLabel = false,
 		strokeWidth = '10px',
 		strokeLinecap = 'round',
 		// Root
@@ -46,8 +47,8 @@
 	const api = $derived(progress.connect(snapshot, send, normalizeProps));
 
 	// Reactive Classes
-	const rxAnimCircle = $derived(api.indeterminate && 'animate-spin');
-	const rxAnimMeter = $derived(api.indeterminate && meterAnimate);
+	const rxAnimCircle = $derived(api.indeterminate ? 'animate-spin' : '');
+	const rxAnimMeter = $derived(api.indeterminate ? meterAnimate : '');
 </script>
 
 <!-- @component A circular progress bar. -->
@@ -74,7 +75,7 @@
 			data-testid="progress-ring-meter"
 		/>
 		<!-- Label -->
-		{#if api.value !== null && !children}
+		{#if api.value !== null && !children && showLabel}
 			<text
 				class="{labelBase} {labelFill} {labelClasses}"
 				x="50%"