+++ sites/next.skeleton.dev/src/content/config.ts
@@ -10,6 +10,8 @@ const docs = defineCollection({
 		srcAlly: z.string().optional(),
 		srcZag: z.string().optional(),
 		showDocsUrl: z.boolean().optional().default(false),
+		pubDate: z.date().optional(),
+		tags: z.array(z.string()).optional(),
 		order: z.number().optional().default(0)
 	})
 });
@@ -19,4 +21,30 @@ const blog = defineCollection({
 	schema: z.object({})
 });
 
-export const collections = { docs, blog };
+const schemas = defineCollection({
+	type: 'data',
+	schema: z.array(
+		z.object({
+			name: z.string(),
+			properties: z.array(
+				z.object({
+					name: z.string(),
+					type: z.string(),
+					typeKind: z.string(),
+					required: z.boolean(),
+					documentation: z.object({
+						text: z.string().nullable(),
+						tags: z.array(
+							z.object({
+								name: z.string(),
+								value: z.string()
+							})
+						)
+					})
+				})
+			)
+		})
+	)
+});
+
+export const collections = { docs, blog, schemas };