+++ packages/skeleton/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @skeletonlabs/skeleton
 
+## 3.0.0-next.9
+
+### Patch Changes
+
+- fix: add `string` type to `contentPath` args ([#2954](https://github.com/skeletonlabs/skeleton/pull/2954))
+
+## 3.0.0-next.8
+
+### Patch Changes
+
+- feature: Added a `contentPath` utility function for tailwind content paths ([#2949](https://github.com/skeletonlabs/skeleton/pull/2949))
+
+## 3.0.0-next.7
+
+### Patch Changes
+
+- Fix conflicting build artifacts ([#2931](https://github.com/skeletonlabs/skeleton/pull/2931))
+
 ## 3.0.0-next.6
 
 ### Minor Changes